# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ATP — At The Park API
  description: |
    REST API powering the At The Park fitness community site (UAE + Oman).
    All write endpoints require a JWT bearer token unless noted.

    Production base URL: `https://atpworldweb-production.up.railway.app/api`
    Versioned alias:     `https://atpworldweb-production.up.railway.app/api/v1`
  version: "1.0"
  contact:
    name: ATP team
    url:  https://atpworldweb-production.up.railway.app
servers:
  - url: https://atpworldweb-production.up.railway.app/api
    description: Production
  - url: https://atpworldweb-production.up.railway.app/api/v1
    description: Production (versioned)
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearer: []
paths:
...
  /community/posts:
    post:
      tags: [Community]
      summary: Create a post
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { content: { type: string }, media: { type: array } } } } }
      responses: { '201': { description: Created } }
  /community/posts/{id}:
    get:
      tags: [Community]
      summary: Get one post (same shape and visibility as the feed)
      security: []
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Post }, '404': { description: Not found or not visible } }
...